+++ src/main/java/cloud/filibuster/junit/server/core/test_executions/ConcreteTestExecution.java
@@ -1,6 +1,7 @@
 package cloud.filibuster.junit.server.core.test_executions;
 
 import cloud.filibuster.dei.DistributedExecutionIndex;
+import cloud.filibuster.junit.server.core.test_execution_reports.TestExecutionReport;
 import org.json.JSONObject;
 
 @SuppressWarnings("Varifier")
